turning into the repo for a paper
by now the BERD pilot application has been reviewed and it will soon be
time to submit aorsf to ROpenSci. Per the ROpenSci ML standard to
show the functionality of aorsf compared to similar packages, I am
setting this repo up as a project to coordinate benchmarking aorsf
versus similar software. Once the project is developed more, I will
convert it to a paper.